requirements.txt
Flask==2.3.3
requests==2.31.0
python-json-logger==2.0.7
python-dotenv==1.0.0
supabase==1.2.0
bcrypt==4.0.1
flask-cors==4.0.0
python-magic-bin==0.4.14